refactor: single-source EvalResult type from core
Move result schemas/types into the zod-only eval-metadata leaf and have the web app import + validate against them instead of a hand-rolled duplicate. Fixes the degenerate inferred type from passthrough().extend().
